requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2.19
djangorestframework==3.14.0
ecdsa==0.18.0
idna==3.4
oauthlib==3.2.2
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
urllib3==2.0.3
